Refactor heat private implementation header
