+++ include/private/rdyheatimpl.h
@@ -1,8 +1,27 @@
 #ifndef RDYHEATIMPL_H
 #define RDYHEATIMPL_H
 
-#include <petscsys.h>
+#include <petscvec.h>
+#include <private/rdyconfigimpl.h>
+#include <private/rdymeshimpl.h>
 #include <rdycore.h>
+
+typedef struct {
+  PetscReal* downwelling_shortwave;
+  PetscReal* downwelling_longwave;
+  PetscReal* wind_speed;
+  PetscReal* air_temperature;
+  PetscReal* specific_humidity;
+} RDyHeatForcing;
+
+struct _RDyHeat {
+  RDyMesh*       mesh;
+  RDyConfig*     config;
+  PetscInt       heat_comp;
+  PetscReal      dt;
+  Vec            star_state;
+  RDyHeatForcing forcing;
+};
 
 PETSC_INTERN PetscErrorCode RDyHeatCreate(RDy);
 PETSC_INTERN PetscErrorCode RDyHeatDestroy(RDy);
